requirements.txt
Django==1.8.3
MySQL-python==1.2.5
argparse==1.2.1
django-appconf==1.0.1
django-compressor==1.5
django-crispy-forms==1.4.0
djangorestframework==3.1.3
python-memcached==1.57
six==1.9.0
wsgiref==0.1.2
